feat: 인증에서 가입으로 전환
