+++ src/store/Ystore.ts
@@ -5,7 +5,7 @@ type ModalState = {
   changeStatus: () => void;
 };
 
-export const useModalStore = create<ModalState>()((set) => ({
+export const useModalStore = create<ModalState>((set) => ({
   modalStatus: false,
   changeStatus: () => set((state) => ({ modalStatus: !state.modalStatus })),
 }));
@@ -15,7 +15,38 @@ type ValidateState = {
   setCertification: (date: string) => void;
 };
 
-export const useValidate = create<ValidateState>()((set) => ({
+export const useValidate = create<ValidateState>((set) => ({
   certification: '',
   setCertification: (data) => set(() => ({ certification: data })),
+}));
+
+type userInfo = {
+  id: string;
+  pw: string;
+  nickName: string;
+};
+
+type useRegisterStoreProps = {
+  isAgreed: boolean;
+  email: string;
+  userInfo: userInfo;
+  setAgree: () => void;
+  setEmail: (emailData: string) => void;
+  setUserInfo: (userData: userInfo) => void;
+};
+
+export const useRegisterStore = create<useRegisterStoreProps>((set) => ({
+  isAgreed: false,
+  email: '',
+  userInfo: {
+    id: '',
+    pw: '',
+    nickName: '',
+  },
+  setAgree: () => set(() => ({ isAgreed: true })),
+  setEmail: (emailData) => set(() => ({ email: emailData })),
+  setUserInfo: (userData) =>
+    set(() => ({
+      userInfo: { ...userData },
+    })),
 }));
